requirements.txt
bcrypt==3.2.0
biopython==1.78
BlazeUtils==0.6.4
cffi==1.14.4
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Materialize==0.2.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
FormEncode==2.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.2.0
Pillow==8.0.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
